Day 4/Hands On/PDFProcessor/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-ai-formrecognizer>=3.3.0
azure-storage-blob>=12.20.0
